refactor(fmt): reuse relative path resolver
